requirements.txt
torch==2.2.2
accelerate==0.30.0
xformers==0.0.25.post1
diffusers==0.27.2
transformers==4.40.0
Pillow
huggingface_hub==0.25.2
peft==0.10.0
numpy==1.24.0
boto3